+++ lib/paper_trail/config.rb
@@ -14,7 +14,8 @@ module PaperTrail
       :object_changes_adapter,
       :serializer,
       :version_limit,
-      :has_paper_trail_defaults
+      :has_paper_trail_defaults,
+      :association_touch_versions
     )
 
     def initialize
@@ -25,6 +26,7 @@ module PaperTrail
       # Variables which affect all threads, whose access is *not* synchronized.
       @serializer = PaperTrail::Serializers::YAML
       @has_paper_trail_defaults = {}
+      @association_touch_versions = true
     end
 
     # Indicates whether PaperTrail is on or off. Default: true.