+++ spec/models/order_spec.rb
@@ -12,4 +12,20 @@ RSpec.describe Order, type: :model, versioning: true do
       expect(customer.versions.count).to(eq(3))
     end
   end
+
+  context "when option: association_touch_versions is set false" do
+    around do |example|
+      PaperTrail.config.association_touch_versions = false
+      example.run
+      PaperTrail.config.association_touch_versions = true
+    end
+
+    it "does not create a version record for association" do
+      customer = Customer.create!
+      order = described_class.create!(customer_id: customer.id)
+      order.reload.update(order_date: Time.now.getlocal)
+
+      expect(customer.versions.count).to(eq(1))
+    end
+  end
 end